==> preprocessing/1_combine_aq_met.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import warnings
from IPython.display import clear_output
warnings.filterwarnings('ignore')

### Data loading

district_data = pd.read_csv('../data/raw/district.csv', dtype='str')
district_data.head(2)

# City id is 1 for Beijing
beijing_districts = district_data[district_data.city_id=='001']['district_id']
beijing_districts

station_data = pd.read_csv('../data/raw/station.csv', dtype='str')

beijing_stations = station_data[station_data.district_id.isin(beijing_districts)]
beijing_stations

beijing_stations.to_csv('tmp_beijing_stations.csv.gz', index=None)

aq_data = pd.read_csv('../data/raw/airquality.csv.gz', dtype='str')
aq_data.head(2)

beijing_aq = aq_data[aq_data.station_id.isin(beijing_stations.station_id)]
beijing_aq.describe()

met_data = pd.read_csv('../data/raw/meteorology.csv.gz', dtype='str')
met_data.head(2)

beijing_met = met_data[met_data.id.isin(beijing_districts)]
beijing_met.head(2)

## Combining

beijing_aq_with_dist = pd.merge(beijing_aq, beijing_stations, )
beijing_aq_with_dist.shape, beijing_aq.shape

beijing_aq_with_dist.head(2)

date_index = pd.date_range('2014-05-01 00:00:00', '2015-04-30 23:00:00', freq='H')
station_index = beijing_aq.station_id.unique()

combo_df = pd.DataFrame(index=pd.MultiIndex.from_product([date_index, station_index], 
                                                         names=['time', 'station_id']))
combo_df.head(2)

beijing_aq['time'] = pd.to_datetime(beijing_aq['time'])
beijing_aq2 = beijing_aq.set_index(['time', 'station_id'])
beijing_aq2.head(2)

### Merging AQ data to global df

for col in beijing_aq2.columns:
    combo_df[col] = beijing_aq2[col]

combo_df

combo_df.to_csv("beijing_aq_all_time_stamps.csv.gz")

### Merging Meteorological data in global df

beijing_met2 = beijing_met.rename(columns={'id': 'district_id'})
beijing_met2['time'] = pd.to_datetime(beijing_met2['time'])
beijing_met2.head(2)
beijing_met_with_station = pd.merge(beijing_met2, beijing_stations).drop(columns=['district_id', 'name_chinese', 
                                                                                  'name_english', 'latitude', 'longitude'])

beijing_met_with_station = beijing_met_with_station.set_index(['time', 'station_id'])
beijing_met_with_station

date_index = pd.date_range('2014-05-01 00:00:00', '2015-04-30 23:00:00', freq='H')
station_index = beijing_aq.station_id.unique()

combo_df = pd.DataFrame(index=pd.MultiIndex.from_product([date_index, station_index], 
                                                         names=['time', 'station_id']))
combo_df.head(2)

for col in beijing_met_with_station.columns:
    combo_df[col] = beijing_met_with_station[col]

combo_df

combo_df.to_csv("beijing_met_all_time_stamps.csv.gz")
