==> preprocessing/3_preprocessing.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import warnings
warnings.filterwarnings('ignore')

beijing_aq = pd.read_csv('beijing_aq_all_time_stamps.csv.gz')
beijing_aq.head(2)

beijing_met = pd.read_csv('beijing_met_all_time_stamps.csv.gz')
beijing_met.head(2)

### Drop pressure because of too much missing data

beijing_met2 = beijing_met.drop(columns=['pressure'])
beijing_met2.head(2)

beijing_met2

### Drop five stations with highest amount of missing data

for station in beijing_met2.station_id.unique():
    tmp_df = beijing_met2[beijing_met2.station_id==station]
    print(station, tmp_df['temperature'].dropna().shape[0]/tmp_df['temperature'].shape[0])

#### Missing data Station IDs: 1009, 1013, 1015, 1020, 1021

beijing_met3 = beijing_met2[~beijing_met2.station_id.isin([1009, 1013, 1015, 1020, 1021])]
beijing_aq2 = beijing_aq[~beijing_aq.station_id.isin([1009, 1013, 1015, 1020, 1021])]

beijing_met2.station_id.unique().shape, beijing_met3.station_id.unique().shape

beijing_met3.to_csv('beijing_met_preprocessed.csv.gz', index=None)
beijing_aq2.to_csv('beijing_aq_preprocessed.csv.gz', index=None)
