==> preprocessing/4_imputing_missing_data.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import warnings
from datetime import datetime
from sklearn.model_selection import ShuffleSplit
from sklearn.metrics import mean_squared_error
from tqdm import tqdm
warnings.filterwarnings('ignore')

beijing_aq = pd.read_csv('beijing_aq_preprocessed.csv.gz')
beijing_aq.head(2)

beijing_met = pd.read_csv('beijing_met_preprocessed.csv.gz')
beijing_met.head(2)

# beijing_met['time'] = pd.to_datetime(beijing_met['time'])
# beijing_aq['time'] = pd.to_datetime(beijing_aq['time'])
# beijing_aq_plus_met2 = beijing_aq_plus_met.sort_values(by=['time', 'station_id'])
# beijing_aq_plus_met2.head(2)

### Asserting number of data entries per station

# for station in beijing_aq_plus_met2.station_id.unique():
#     tmp_df = beijing_aq_plus_met2[beijing_aq_plus_met2.station_id==station]
#     assert tmp_df.shape == (8760, 13)
methods = ['linear','slinear','quadratic','cubic','spline','akima']

### Method 1 : Avg of RMSE

for cols in ['temperature','humidity','wind_speed']:
    
    alinear_loss,aslinear_loss,aquadratic_loss,acubic_loss,aspline_loss,aakima_loss = [],[],[],[],[],[]
    
    for station in tqdm(beijing_met.station_id.unique()):
        
        linear_loss,slinear_loss,quadratic_loss,cubic_loss,spline_loss,akima_loss = [],[],[],[],[],[]
        tmp_df = beijing_met[beijing_met.station_id==station].reset_index(drop=True)
        assert tmp_df.shape == (8760, len(beijing_met.columns))
        
        tmp_df = tmp_df.iloc[:,[beijing_met.columns.get_loc('time'),beijing_met.columns.get_loc(cols)]]
        
        dates = pd.to_datetime([datetime.fromisoformat(tmp_df.iloc[i,tmp_df.columns.get_loc('time')]) for i in range(len(tmp_df))])
        time_differences = ((dates - dates[0]).total_seconds() // 3600)
        tmp_df['delta_t'] = time_differences
        tmp_df = tmp_df.drop(columns='time')
        
        tmp_df_notmissing = tmp_df.dropna(subset=[cols]).reset_index(drop=True)
        indices = ShuffleSplit(n_splits=5, test_size=.2, random_state=0).split(tmp_df_notmissing)
        
        for i in indices:
            train_indices = list(i[0])
            test_indices = list(i[1])
            train_indices.sort(),test_indices.sort()
   
            xtest = tmp_df_notmissing.iloc[test_indices,:]
            xtest = xtest.set_index('delta_t')
            xtrain = tmp_df_notmissing.copy()
            xtrain.iloc[test_indices,xtrain.columns.get_loc(cols)] = np.nan
            xtrain = xtrain.set_index('delta_t')
            
            interpolate_linear = xtrain.interpolate(method='linear')
            interpolate_linear = interpolate_linear.ffill()
            interpolate_linear = interpolate_linear.bfill()
            
            interpolate_slinear = xtrain.interpolate(method='slinear')
            interpolate_slinear = interpolate_slinear.ffill()
            interpolate_slinear = interpolate_slinear.bfill()
            
            interpolate_quadratic = xtrain.interpolate(method='quadratic')
            interpolate_quadratic = interpolate_quadratic.ffill()
            interpolate_quadratic = interpolate_quadratic.bfill()
            
            interpolate_cubic = xtrain.interpolate(method='cubic')
            interpolate_cubic = interpolate_cubic.ffill()
            interpolate_cubic = interpolate_cubic.bfill()
            
            interpolate_spline = xtrain.interpolate(method='spline',order=3)
            interpolate_spline = interpolate_spline.ffill()
            interpolate_spline = interpolate_spline.bfill()
            
            interpolate_akima = xtrain.interpolate(method='akima')
            interpolate_akima = interpolate_akima.ffill()
            interpolate_akima = interpolate_akima.bfill()

            mse_linear = mean_squared_error(xtest,interpolate_linear.iloc[test_indices,0],squared=False)
            mse_slinear = mean_squared_error(xtest,interpolate_slinear.iloc[test_indices,0],squared=False)
            mse_quadratic = mean_squared_error(xtest,interpolate_quadratic.iloc[test_indices,0],squared=False)
            mse_cubic = mean_squared_error(xtest,interpolate_cubic.iloc[test_indices,0],squared=False)
            mse_spline = mean_squared_error(xtest,interpolate_spline.iloc[test_indices,0],squared=False)
            mse_akima = mean_squared_error(xtest,interpolate_akima.iloc[test_indices,0],squared=False)
            
            linear_loss.append(mse_linear),slinear_loss.append(mse_slinear),quadratic_loss.append(mse_quadratic),cubic_loss.append(mse_cubic)
            spline_loss.append(mse_spline),akima_loss.append(mse_akima)
            
        alinear_loss.append(np.mean(mse_linear)),aslinear_loss.append(np.mean(mse_slinear)),aquadratic_loss.append(np.mean(mse_quadratic)),acubic_loss.append(np.mean(mse_cubic))
        aspline_loss.append(np.mean(mse_spline)),aakima_loss.append(np.mean(mse_akima))
        
    print(np.mean(alinear_loss),np.mean(aslinear_loss),np.mean(aquadratic_loss),np.mean(acubic_loss),np.mean(aspline_loss),np.mean(aakima_loss))
    method = methods[np.argmin(np.array([np.mean(alinear_loss),np.mean(aslinear_loss),np.mean(aquadratic_loss),np.mean(acubic_loss),np.mean(aspline_loss),np.mean(aakima_loss)]))]
    print(cols,method)

for cols in ['PM25_Concentration']:
    
    alinear_loss,aslinear_loss,aquadratic_loss,acubic_loss,aspline_loss,aakima_loss = [],[],[],[],[],[]
    
    for station in tqdm(beijing_aq.station_id.unique()):
        
        linear_loss,slinear_loss,quadratic_loss,cubic_loss,spline_loss,akima_loss = [],[],[],[],[],[]
        tmp_df = beijing_aq[beijing_aq.station_id==station].reset_index(drop=True)
        assert tmp_df.shape == (8760, len(beijing_aq.columns))
        
        tmp_df = tmp_df.iloc[:,[beijing_aq.columns.get_loc('time'),beijing_aq.columns.get_loc(cols)]]
        
        dates = pd.to_datetime([datetime.fromisoformat(tmp_df.iloc[i,tmp_df.columns.get_loc('time')]) for i in range(len(tmp_df))])
        time_differences = ((dates - dates[0]).total_seconds() // 3600)
        tmp_df['delta_t'] = time_differences
        tmp_df = tmp_df.drop(columns='time')
        
        tmp_df_notmissing = tmp_df.dropna(subset=[cols]).reset_index(drop=True)
        indices = ShuffleSplit(n_splits=5, test_size=.2, random_state=0).split(tmp_df_notmissing)
        
        for i in indices:
            train_indices = list(i[0])
            test_indices = list(i[1])
            train_indices.sort(),test_indices.sort()
   
            xtest = tmp_df_notmissing.iloc[test_indices,:]
            xtest = xtest.set_index('delta_t')
            xtrain = tmp_df_notmissing.copy()
            xtrain.iloc[test_indices,xtrain.columns.get_loc(cols)] = np.nan
            xtrain = xtrain.set_index('delta_t')
            
            interpolate_linear = xtrain.interpolate(method='linear')
            interpolate_linear = interpolate_linear.ffill()
            interpolate_linear = interpolate_linear.bfill()
            
            interpolate_slinear = xtrain.interpolate(method='slinear')
            interpolate_slinear = interpolate_slinear.ffill()
            interpolate_slinear = interpolate_slinear.bfill()
            
            interpolate_quadratic = xtrain.interpolate(method='quadratic')
            interpolate_quadratic = interpolate_quadratic.ffill()
            interpolate_quadratic = interpolate_quadratic.bfill()
            
            interpolate_cubic = xtrain.interpolate(method='cubic')
            interpolate_cubic = interpolate_cubic.ffill()
            interpolate_cubic = interpolate_cubic.bfill()
            
            interpolate_spline = xtrain.interpolate(method='spline',order=3)
            interpolate_spline = interpolate_spline.ffill()
            interpolate_spline = interpolate_spline.bfill()
            
            interpolate_akima = xtrain.interpolate(method='akima')
            interpolate_akima = interpolate_akima.ffill()
            interpolate_akima = interpolate_akima.bfill()

            mse_linear = mean_squared_error(xtest,interpolate_linear.iloc[test_indices,0],squared=False)
            mse_slinear = mean_squared_error(xtest,interpolate_slinear.iloc[test_indices,0],squared=False)
            mse_quadratic = mean_squared_error(xtest,interpolate_quadratic.iloc[test_indices,0],squared=False)
            mse_cubic = mean_squared_error(xtest,interpolate_cubic.iloc[test_indices,0],squared=False)
            mse_spline = mean_squared_error(xtest,interpolate_spline.iloc[test_indices,0],squared=False)
            mse_akima = mean_squared_error(xtest,interpolate_akima.iloc[test_indices,0],squared=False)
            
            linear_loss.append(mse_linear),slinear_loss.append(mse_slinear),quadratic_loss.append(mse_quadratic),cubic_loss.append(mse_cubic)
            spline_loss.append(mse_spline),akima_loss.append(mse_akima)
            
        alinear_loss.append(np.mean(mse_linear)),aslinear_loss.append(np.mean(mse_slinear)),aquadratic_loss.append(np.mean(mse_quadratic)),acubic_loss.append(np.mean(mse_cubic))
        aspline_loss.append(np.mean(mse_spline)),aakima_loss.append(np.mean(mse_akima))
        
    print(np.mean(alinear_loss),np.mean(aslinear_loss),np.mean(aquadratic_loss),np.mean(acubic_loss),np.mean(aspline_loss),np.mean(aakima_loss))
    method = methods[np.argmin(np.array([np.mean(alinear_loss),np.mean(aslinear_loss),np.mean(aquadratic_loss),np.mean(acubic_loss),np.mean(aspline_loss),np.mean(aakima_loss)]))]
    print(cols,method)

### Method 2 : Flatten and take RMSE

for cols in ['temperature','humidity','wind_speed']:
    
    testing,ind = [],[]
    linear_loss,slinear_loss,quadratic_loss,cubic_loss,spline_loss,akima_loss = [],[],[],[],[],[]
    
    for station in tqdm(beijing_met.station_id.unique()):
        
        
        tmp_df = beijing_met[beijing_met.station_id==station].reset_index(drop=True)
        assert tmp_df.shape == (8760, len(beijing_met.columns))
        
        tmp_df = tmp_df.iloc[:,[beijing_met.columns.get_loc('time'),beijing_met.columns.get_loc(cols)]]
        
        dates = pd.to_datetime([datetime.fromisoformat(tmp_df.iloc[i,tmp_df.columns.get_loc('time')]) for i in range(len(tmp_df))])
        time_differences = ((dates - dates[0]).total_seconds() // 3600)
        tmp_df['delta_t'] = time_differences
        tmp_df = tmp_df.drop(columns='time')
        
        tmp_df_notmissing = tmp_df.dropna(subset=[cols]).reset_index(drop=True)
        indices = ShuffleSplit(n_splits=5, test_size=.2, random_state=0).split(tmp_df_notmissing)
        
        for i in indices:
            train_indices = list(i[0])
            test_indices = list(i[1])
            train_indices.sort(),test_indices.sort()
   
            xtest = tmp_df_notmissing.iloc[test_indices,:]
            xtest = xtest.set_index('delta_t')
            xtrain = tmp_df_notmissing.copy()
            xtrain.iloc[test_indices,xtrain.columns.get_loc(cols)] = np.nan
            xtrain = xtrain.set_index('delta_t')
            
            interpolate_linear = xtrain.interpolate(method='linear')
            interpolate_linear = interpolate_linear.ffill()
            interpolate_linear = interpolate_linear.bfill()
            
            interpolate_slinear = xtrain.interpolate(method='slinear')
            interpolate_slinear = interpolate_slinear.ffill()
            interpolate_slinear = interpolate_slinear.bfill()
            
            interpolate_quadratic = xtrain.interpolate(method='quadratic')
            interpolate_quadratic = interpolate_quadratic.ffill()
            interpolate_quadratic = interpolate_quadratic.bfill()
            
            interpolate_cubic = xtrain.interpolate(method='cubic')
            interpolate_cubic = interpolate_cubic.ffill()
            interpolate_cubic = interpolate_cubic.bfill()
            
            interpolate_spline = xtrain.interpolate(method='spline',order=3)
            interpolate_spline = interpolate_spline.ffill()
            interpolate_spline = interpolate_spline.bfill()
            
            interpolate_akima = xtrain.interpolate(method='akima')
            interpolate_akima = interpolate_akima.ffill()
            interpolate_akima = interpolate_akima.bfill()

            
            linear_loss.append(interpolate_linear.iloc[test_indices,0]),slinear_loss.append(interpolate_slinear.iloc[test_indices,0]),quadratic_loss.append(interpolate_quadratic.iloc[test_indices,0]),cubic_loss.append(interpolate_cubic.iloc[test_indices,0])
            spline_loss.append(interpolate_spline.iloc[test_indices,0]),akima_loss.append(interpolate_akima.iloc[test_indices,0])
            testing.append(xtest)
    
    interpolate_linear = list(np.concatenate(linear_loss).flat)
    interpolate_slinear = list(np.concatenate(slinear_loss).flat)
    interpolate_quadratic = list(np.concatenate(quadratic_loss).flat)
    interpolate_cubic = list(np.concatenate(cubic_loss).flat)
    interpolate_spline = list(np.concatenate(spline_loss).flat)
    interpolate_akima = list(np.concatenate(akima_loss).flat)
    
    xtest = list(np.concatenate(testing).flat)
#     print(xtest)
    mse_linear = mean_squared_error(xtest,interpolate_linear,squared=False)
    mse_slinear = mean_squared_error(xtest,interpolate_slinear,squared=False)
    mse_quadratic = mean_squared_error(xtest,interpolate_quadratic,squared=False)
    mse_cubic = mean_squared_error(xtest,interpolate_cubic,squared=False)
    mse_spline = mean_squared_error(xtest,interpolate_spline,squared=False)
    mse_akima = mean_squared_error(xtest,interpolate_akima,squared=False)
            
    method = methods[np.argmin(np.array([mse_linear,mse_slinear,mse_quadratic,mse_cubic,mse_spline,mse_akima]))]
    print(cols,method)

for cols in ['PM25_Concentration']:
    
    testing,ind = [],[]
    linear_loss,slinear_loss,quadratic_loss,cubic_loss,spline_loss,akima_loss = [],[],[],[],[],[]
    
    for station in tqdm(beijing_aq.station_id.unique()):
        
        
        tmp_df = beijing_aq[beijing_aq.station_id==station].reset_index(drop=True)
        assert tmp_df.shape == (8760, len(beijing_aq.columns))
        
        tmp_df = tmp_df.iloc[:,[beijing_aq.columns.get_loc('time'),beijing_aq.columns.get_loc(cols)]]
        
        dates = pd.to_datetime([datetime.fromisoformat(tmp_df.iloc[i,tmp_df.columns.get_loc('time')]) for i in range(len(tmp_df))])
        time_differences = ((dates - dates[0]).total_seconds() // 3600)
        tmp_df['delta_t'] = time_differences
        tmp_df = tmp_df.drop(columns='time')
        
        tmp_df_notmissing = tmp_df.dropna(subset=[cols]).reset_index(drop=True)
        indices = ShuffleSplit(n_splits=5, test_size=.2, random_state=0).split(tmp_df_notmissing)
        
        for i in indices:
            train_indices = list(i[0])
            test_indices = list(i[1])
            train_indices.sort(),test_indices.sort()
   
            xtest = tmp_df_notmissing.iloc[test_indices,:]
            xtest = xtest.set_index('delta_t')
            xtrain = tmp_df_notmissing.copy()
            xtrain.iloc[test_indices,xtrain.columns.get_loc(cols)] = np.nan
            xtrain = xtrain.set_index('delta_t')
            
            interpolate_linear = xtrain.interpolate(method='linear')
            interpolate_linear = interpolate_linear.ffill()
            interpolate_linear = interpolate_linear.bfill()
            
            interpolate_slinear = xtrain.interpolate(method='slinear')
            interpolate_slinear = interpolate_slinear.ffill()
            interpolate_slinear = interpolate_slinear.bfill()
            
            interpolate_quadratic = xtrain.interpolate(method='quadratic')
            interpolate_quadratic = interpolate_quadratic.ffill()
            interpolate_quadratic = interpolate_quadratic.bfill()
            
            interpolate_cubic = xtrain.interpolate(method='cubic')
            interpolate_cubic = interpolate_cubic.ffill()
            interpolate_cubic = interpolate_cubic.bfill()
            
            interpolate_spline = xtrain.interpolate(method='spline',order=3)
            interpolate_spline = interpolate_spline.ffill()
            interpolate_spline = interpolate_spline.bfill()
            
            interpolate_akima = xtrain.interpolate(method='akima')
            interpolate_akima = interpolate_akima.ffill()
            interpolate_akima = interpolate_akima.bfill()

            
            linear_loss.append(interpolate_linear.iloc[test_indices,0]),slinear_loss.append(interpolate_slinear.iloc[test_indices,0]),quadratic_loss.append(interpolate_quadratic.iloc[test_indices,0]),cubic_loss.append(interpolate_cubic.iloc[test_indices,0])
            spline_loss.append(interpolate_spline.iloc[test_indices,0]),akima_loss.append(interpolate_akima.iloc[test_indices,0])
            testing.append(xtest)
    
    interpolate_linear = list(np.concatenate(linear_loss).flat)
    interpolate_slinear = list(np.concatenate(slinear_loss).flat)
    interpolate_quadratic = list(np.concatenate(quadratic_loss).flat)
    interpolate_cubic = list(np.concatenate(cubic_loss).flat)
    interpolate_spline = list(np.concatenate(spline_loss).flat)
    interpolate_akima = list(np.concatenate(akima_loss).flat)
    
    xtest = list(np.concatenate(testing).flat)
#     print(xtest)
    mse_linear = mean_squared_error(xtest,interpolate_linear,squared=False)
    mse_slinear = mean_squared_error(xtest,interpolate_slinear,squared=False)
    mse_quadratic = mean_squared_error(xtest,interpolate_quadratic,squared=False)
    mse_cubic = mean_squared_error(xtest,interpolate_cubic,squared=False)
    mse_spline = mean_squared_error(xtest,interpolate_spline,squared=False)
    mse_akima = mean_squared_error(xtest,interpolate_akima,squared=False)
            
    method = methods[np.argmin(np.array([mse_linear,mse_slinear,mse_quadratic,mse_cubic,mse_spline,mse_akima]))]
    print(cols,method)

station_wise_met_df,station_wise_aq_df = [],[]
best_methods = {'PM25_Concentration': 'akima', 
                'temperature': 'slinear',
                'humidity': 'slinear', 
                'wind_speed': 'slinear',
                'weather': 'nearest',
                'wind_direction': 'nearest'}
for station in beijing_met.station_id.unique():
    tmp_df = beijing_met[beijing_met.station_id == station]
    for cols in beijing_met.columns:
        try:
            tmp_df[cols] = tmp_df[cols].interpolate(best_methods[cols]).ffill().bfill()
        except:
            continue
    station_wise_met_df.append(tmp_df)
    tmp_df = beijing_aq[beijing_aq.station_id == station]
    for cols in beijing_aq.columns:
        try:
            tmp_df[cols] = tmp_df[cols].interpolate(best_methods[cols]).ffill().bfill()
        except:
            continue
    station_wise_aq_df.append(tmp_df)
    
filled_df_met = pd.concat(station_wise_met_df)
filled_df_aq = pd.concat(station_wise_aq_df)

merged_aq_met = pd.merge(filled_df_aq,filled_df_met)
display(merged_aq_met)
